pyproject.toml
[tool.poetry]
name = "vlm-manipulation"
version = "0.1.0"
description = "Manipulation in the simulation and real-world environment. GPT-4 and G-DINO based reasoning is utilized for commonsense image reasoning."
authors = ["SeungyounShin <ssy10011218@gmail.com>", "Hyeonseong Kim <hyeonseong-kim@korea.ac.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.30.3"
pillow = "^10.3.0"
termcolor = "^2.4.0"
prettytable = "^3.10.0"
rich = "^13.7.1"
transformers = "^4.41.1"
torch = "^2.0.0"
torchvision = "^0.15.0"
torchaudio = "^2.0.0"
instructor = "^1.3.1"
pydantic = "^2.7.1"
matplotlib = "^3.9.0"
scipy = "^1.13.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
